gmail_lookup_service/requirements.txt
fastapi==0.115.5
jinja2==3.1.4
uvicorn==0.30.6
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
sounddevice==0.4.7
soundfile==0.12.1
